demos/requirements.txt
# 核心机器学习和数据科学
numpy==1.26.4
pandas==2.0.3
scipy==1.13.0
scikit-learn==1.3.2

# PyTorch相关
-f https://data.pyg.org/whl/torch-2.1.1+cpu.html
torch==2.1.1
torchvision==0.16.1
torch-scatter==2.1.2
torch-sparse==0.6.18
torch_geometric==2.4.0
dgl==1.1.3

# 生物信息学和分析
scanpy==1.10.1
anndata==0.10.8
scib==1.1.5
gseapy==1.1.9
mudata==0.2.3

# 可视化
matplotlib==3.8.0
seaborn==0.13.2
plotly==6.1.1

# API和Web服务
fastapi==0.115.12
uvicorn==0.34.3
wandb==0.16.3
requests==2.31.0
python-multipart==0.0.20

# 实用工具
tqdm==4.66.2
networkx==3.3
POT==0.9.4
PyYAML==6.0.2
omegaconf==2.3.0
openpyxl==3.1.2
